### 2018-11-18 #1074 @kthyng
Hi @fgnievinski and @mvdh7: are you willing and able to review this JOSS submission? You can read more about the process [here](https://joss.readthedocs.io/en/latest/review_criteria.html). We normally ask for reviews in about 2 weeks but given there is a holiday in the US next week, I would like reviews in the next 3 weeks.
