### 2018-11-21 deb6617
Update paper.md
