+++ paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'An MATLAB toolbox to detect and analyze marine heatwaves'
+title: 'A MATLAB toolbox to detect and analyze marine heatwaves'
 tags:
 - MATLAB
 - heatwaves
