+++ README.md
@@ -148,7 +148,7 @@ Contributing to **`m_mhw`**
 
 To contribute to the package please follow the guidelines [here](https://github.com/ZijieZhao/see/blob/master/Contributing_to_mmhw.md).
 
-Please use this [link](https://github.com/ZijieZhao/see/issues) to report any bugs found.
+Please use this [link](https://github.com/ZijieZhaoMMHW/m_mhw1.0/issues) to report any bugs found.
 
 References
 ----------
