### 2018-12-03 #1074 @kthyng
Hey @chadagreene! Are you able and willing to review this submission in this next two weeks? I think you'd be a good fit. You can get more info on reviewing [here](https://joss.readthedocs.io/en/latest/review_criteria.html).
