### 2018-12-13 ddfe405
Update paper.md
