### 2018-12-13 f1ccbb2
Update detect.m
