+++ detect.m
@@ -1,29 +1,29 @@
-function [MHW,mclim,m90,mhw_ts]=detect(temp,cli_start,cli_end,data_start,mhw_start,mhw_end,varargin)
+function [MHW,mclim,m90,mhw_ts]=detect(temp,time,cli_start,cli_end,mhw_start,mhw_end,varargin)
 %detect - Detecting spatial MHW/MCS
 %  Syntax
 %
-%  [MHW]=detect(temp,cli_start,cli_end,data_start,mhw_start,mhw_end)
-%  [MHW,mclim,m90,mhw_ts]=detect(temp,cli_start,cli_end,data_start,mhw_start,mhw_end);
-%  [MHW,mclim,m90,mhw_ts]=detect(temp,cli_start,cli_end,data_start,mhw_start,mhw_end,'Event','MCS','Threshold',0.1);
+%  [MHW]=detect(temp,time,cli_start,cli_end,mhw_start,mhw_end)
+%  [MHW,mclim,m90,mhw_ts]=detect(temp,time,cli_start,cli_end,mhw_start,mhw_end);
+%  [MHW,mclim,m90,mhw_ts]=detect(temp,time,cli_start,cli_end,mhw_start,mhw_end,'Event','MCS','Threshold',0.1);
 %
 %  Description
 %
-%  [MHW]=detect(temp,cli_start,cli_end,data_start,mhw_start,mhw_end) returns
+%  [MHW]=detect(temp,time,cli_start,cli_end,mhw_start,mhw_end) returns
 %  all detected MHW events for the m-by-n-by-t matrix TEMP starting in the
 %  year DATA_START. m, n and t separately indicate two spatial dimensions
 %  (m and n) and one temporal dimension (t). Climatologies used to
-%  determine events are calculated based on TEMP from year CLI_START to
-%  year CLI_END. MHW is a table where each row corresponds to a particular
-%  event during year MHW_START to year MHW_END and each column indicates a metric.
+%  determine events are calculated based on TEMP from CLI_START to
+%  CLI_END. MHW is a table where each row corresponds to a particular
+%  event during MHW_START to MHW_END and each column indicates a metric.
 %
-%  [MHW,mclim,m90,mhw_ts]=detect(temp,cli_start,cli_end,data_start,mhw_start,mhw_end)
+%  [MHW,mclim,m90,mhw_ts]=detect(temp,time,cli_start,cli_end,mhw_start,mhw_end)
 %  also return the spatial climatology MCLIM (m-by-n-by-366) and threshold
 %  M90 (m-by-n-by-366) to calculate MHW events and resultant MHW time series
 %  MHW_TS (m-by-n-by-t). In the condition that there is no missing value in
 %  data TEMP, NaN in all outputs indicates lands and 0 in MHW_TS indicates
 %  the corresponding day in that grid is not in a MHW event.
 %
-%  [MHW,mclim,m90,mhw_ts]=detect(temp,cli_start,cli_end,data_start,mhw_start,mhw_end,'Event','MCS','Threshold',0.1)
+%  [MHW,mclim,m90,mhw_ts]=detect(temp,time,cli_start,cli_end,mhw_start,mhw_end,'Event','MCS','Threshold',0.1)
 %  returns MCS events based on 10th percentile threshold.
 %
 %  Input Arguments
@@ -32,19 +32,22 @@ function [MHW,mclim,m90,mhw_ts]=detect(temp,cli_start,cli_end,data_start,mhw_sta
 %   m-by-n-by-t matrix. m and n separately indicate two spatial dimensions
 %   and t indicates temporal dimension. 
 %
-%   cli_start - A numeric value indicating the start year for the period
+%   time - datenum(start_year,start_month,start_day):datenum(end_year,
+%   end_month,end_day)
+%
[…]
