### 2018-12-13 89a9ca4
Update paper.md
