### 2018-12-13 f66dbe6
Update detect.m
