+++ Changing_History.md
@@ -1,7 +1,0 @@
-Changing History
-==================================================================
-
-2018-12-14
--------------
-
-Updating the algorithm to calculate the climatologies and thresholds with help from Maxime Marin (<Maxime.Marin@csiro.au>).
