### 2018-12-20 8f9231d
Delete LICENCE.txt
