### 2018-12-20 dad1ad2
Delete an_example.m
