+++ an_example.m
@@ -1,448 +1,0 @@
-%% An example about how to apply m_mhw to real-world data
-% Here we provide an example about how to apply m_mhw to real-world data.
-
-%% 1. Loading data
-
-% Load NOAA OI SST V2 data
-sst_full=NaN(32,32,datenum(2016,12,31)-datenum(1982,1,1)+1);
-for i=1982:2016;
-    file_here=['sst_' num2str(i)];
-    load(file_here);
-    eval(['data_here=sst_' num2str(i) ';'])
-    sst_full(:,:,(datenum(i,1,1):datenum(i,12,31))-datenum(1982,1,1)+1)=data_here;
-end
-
-% This data includes SST in [147-155E, 45-37S] in resolution of 0.25 from
-% 1982 to 2016.
-
-load('lon_and_lat');
-size(sst_full); %size of data
-datenum(2016,12,31)-datenum(1982,1,1)+1 % The temporal length is 35 years.
-
-%% 2. Detecting MHWs and MCSs
-
-% Here we detect marine heatwaves off eastern Tasmania based on the
-% traditional definition of MHWs (Hobday et al. 2016). We detected MHWs
-% during 1993 to 2016 for climatologies and thresholds in 1982 to 2005.
-
-[MHW,mclim,m90,mhw_ts]=detect(sst_full,datenum(1982,1,1):datenum(2016,12,31),datenum(1982,1,1),datenum(2005,12,31),datenum(1993,1,1),datenum(2016,12,31));; %take about 30 seconds.
-
-% Additionally, we also detect MCSs during 1982 to 2005 based on the same
-% climatologies. 
-
-[MCS,~,m10,mcs_ts]=detect(sst_full,datenum(1982,1,1):datenum(2016,12,31),datenum(1982,1,1),datenum(2005,12,31),datenum(1993,1,1),datenum(2016,12,31),'Event','MCS','Threshold',0.1);
-
-% Have a look of these two data.
-MHW(1:5,:);
-MCS(1:5,:);
-
-% You could see that the properties `mhw_onset` and `mhw_end` are in a
-% strange format. This is due to the fact that they are originally
-% constructed in numeric format. We could change it to date format by
-% following steps.
-
-datevec(num2str(MHW{1:5,:}),'yyyymmdd') % vector
-datestr(datevec(num2str(MHW{1:5,:}),'yyyymmdd')) % string
-datenum(num2str(MHW{1:5,:}),'yyyymmdd') % number
-
-%% 3. Visualizing MHW/MCS time series
-
-% Have a look of MHW events in grid (1,2) from Sep 2015 to Apr 2016
-figure('pos',[10 10 1000 1000]);
-event_line(sst_full,MHW,mclim,m90,[1 2],1982,[2015 9 1],[2016 5 1]);
-
-% Have a look of MCS events in grid (1,2) during 1994
-figure('pos',[10 10 1000 1000]);
-event_line(sst_full,MCS,mclim,m10,[1 2],1982,[1994 1 1],[1994 12 31],'Event','MCS','Color',[0.5 0.5 1]);
-
-%% 4. Mean states and trends
-
-% Please note that this section requires the toolbox m_map
-
-% Now we would like to know the mean states and annual trends of MHW
-% frequency, i.e. how many MHW events would be detected per year and how it
-% changes with time.
-
[…]
