### 2018-12-20 cbc1d45
Delete paper.bib
