+++ paper.bib
@@ -1,149 +1,0 @@
-@article{garrabou2009mass,
-  title={Mass mortality in Northwestern Mediterranean rocky benthic communities: effects of the 2003 heat wave},
-  author={Garrabou, Joaquim and Coma, Rafael and Bensoussan, Nathaniel and Bally, Marc and Chevaldonn{\'e}, Pierre and Cigliano, M and D{\'\i}az, David and Harmelin, Jean-Georges and Gambi, Mar{\'\i}a Cristina and Kersting, DK and others},
-  journal={Global change biology},
-  volume={15},
-  number={5},
-  pages={1090--1103},
-  year={2009},
-  publisher={Wiley Online Library}
-}
-
-@article{hansen2006global,
-  title={Global temperature change},
-  author={Hansen, James and Sato, Makiko and Ruedy, Reto and Lo, Ken and Lea, David W and Medina-Elizade, Martin},
-  journal={Proceedings of the National Academy of Sciences},
-  volume={103},
-  number={39},
-  pages={14288--14293},
-  year={2006},
-  publisher={National Acad Sciences}
-}
-
-@article{hobday2014identification,
-  title={Identification of global marine hotspots: sentinels for change and vanguards for adaptation action},
-  author={Hobday, Alistair J and Pecl, Gretta T},
-  journal={Reviews in Fish Biology and Fisheries},
-  volume={24},
-  number={2},
-  pages={415--425},
-  year={2014},
-  publisher={Springer}
-}
-
-@article{hobday2016hierarchical,
-  title={A hierarchical approach to defining marine heatwaves},
-  author={Hobday, Alistair J and Alexander, Lisa V and Perkins, Sarah E and Smale, Dan A and Straub, Sandra C and Oliver, Eric CJ and Benthuysen, Jessica A and Burrows, Michael T and Donat, Markus G and Feng, Ming and others},
-  journal={Progress in Oceanography},
-  volume={141},
-  pages={227--238},
-  year={2016},
-  publisher={Elsevier}
-}
-
-@book{pachauri2014climate,
-  title={Climate change 2014: synthesis report. Contribution of Working Groups I, II and III to the fifth assessment report of the Intergovernmental Panel on Climate Change},
-  author={Pachauri, Rajendra K and Allen, Myles R and Barros, Vicente R and Broome, John and Cramer, Wolfgang and Christ, Renate and Church, John A and Clarke, Leon and Dahe, Qin and Dasgupta, Purnamita and others},
-  year={2014},
-  publisher={IPCC}
-}
-
-@article{mills2013fisheries,
-  title={Fisheries management in a changing climate: lessons from the 2012 ocean heat wave in the Northwest Atlantic},
-  author={Mills, Katherine E and Pershing, Andrew J and Brown, Curtis J and Chen, Yong and Chiang, Fu-Sung and Holland, Daniel S and Lehuta, Sigrid and Nye, Janet A and Sun, Jenny C and Thomas, Andrew C and others},
-  journal={Oceanography},
-  volume={26},
-  number={2},
-  pages={191--195},
-  year={2013},
-  publisher={JSTOR}
-}
-
-@misc{new2011four,
-  title={Four degrees and beyond: the potential for a global temperature increase of four degrees and its implications},
-  author={New, Mark and Liverman, Diana and Schroder, Heike and Anderson, Kevin},
-  year={2011},
-  publisher={The Royal Society}
-}
-
-@article{oliver2017unprecedented,
-  title={The unprecedented 2015/16 Tasman Sea marine heatwave},
[…]
