### 2018-12-20 f85f8e0
Delete paper.md
