### 2018-12-20 88f902e
Delete Contributing_to_mmhw.md
