+++ Contributing_to_mmhw.md
@@ -1,22 +1,0 @@
-Contributing to **`m_mhw`**
-==================================================================
-
-Please see the guidelines below for contributions to the **`m_mhw`** package.
-
-Typos and grammatical errors
-----------------
-
-Corrections to typos and grammatical errors may be made directly through the GitHub web interface. Any changes should be made within the .m source file.
-
-Filing an issue
-----------------
-
-Should one want to propose a change or new feature, please file an issue through the GitHub interface. Bug reports specifically may be made [here](https://github.com/ZijieZhao/see/issues). When proposing a change/new feature, or reporting a bug, please illustrate the issue with a minimal reproducible example, including texts, codes and resultant outputs.
-
-Code of Conduct
-----------------
-
-Please note that this project is released with a [Contributor Code of Conduct](https://github.com/ZijieZhao/see/blob/master/Code_of_Conduct.md). By participating in this project you agree to abide by its terms.
-
-
-
