### 2018-12-20 904aa6e
Delete compare_M_R.md
