### 2018-12-20 c55a6a8
Delete an_example.md
