+++ docs/LICENCE.txt
@@ -1,0 +1,2 @@
+YEAR: 2018
+COPYRIGHT HOLDER: Zijie Zhao
