### 2018-12-20 0cc0fa8
Update Contributing_to_mmhw.md
