### 2019-01-08 ZijieZhaoMMHW/m_mhw1.0#1 @mvdh7
Non-OSI license
[For JOSS review](https://github.com/openjournals/joss-reviews/issues/1124)

The license file simply states a name and a year, and does not appear to refer to any from the [OSI approved list](https://opensource.org/licenses/alphabetical) as required by JOSS.
