### 2019-01-08 ZijieZhaoMMHW/m_mhw1.0#2 @mvdh7
No DOIs in paper
[For JOSS review](https://github.com/openjournals/joss-reviews/issues/1124)

There are no DOIs for citations in the paper or its .bib file; these are required by JOSS.
