### 2019-01-08 ZijieZhaoMMHW/m_mhw1.0#3 @mvdh7
Function inputs documentation
[For JOSS review](https://github.com/openjournals/joss-reviews/issues/1124)

Documentation on the functions’ API in the readme includes a description of their outputs but nothing on inputs. Inputs are described in code comments within the functions themselves, but it would be helpful to have this info in the readme too.
