### 2019-01-08 ZijieZhaoMMHW/m_mhw1.0#4 @mvdh7
m_map required for testing
[For JOSS review](https://github.com/openjournals/joss-reviews/issues/1124)

The separate package m_map is required to run the example scripts but no information is provided on how to obtain and install this. Consequently I cannot fully test the functional claims of the software.

m_map is also not listed as a requirement in the documentation. If you consider the example scripts to be not part of the software under review then that could perhaps be justified; m_map is not needed to actually execute the main functions. But as the example scripts are the only straightforward way to test the functionality, I'm not sure that that applies in this case.
