### 2019-01-08 ZijieZhaoMMHW/m_mhw1.0#5 @mvdh7
MATLAB warning flags
[For JOSS review](https://github.com/openjournals/joss-reviews/issues/1124)

The MATLAB code for the functions themselves contains many warning messages indicating inefficient or unnecessary constructs. I strongly suggest the authors work through and eliminate these, to improve code legibility, and in some cases, its performance.
