### 2019-01-08 ZijieZhaoMMHW/m_mhw1.0#6 @mvdh7
Results inconsistencies with other implementations
[For JOSS review](https://github.com/openjournals/joss-reviews/issues/1124)

The noted differences in results between the different implementations (MATLAB, Python, R) are worrisome. It's said that this is caused by differences in how the different languages evaluate percentile thresholds, but surely the code could be adapted to enforce consistent calculations? If not, can you demonstrate that this is the **only** difference between the implementations, and thus prove that no other discrepancies are hiding behind it?
