### 2019-01-08 #1124 @mvdh7
Thanks @kthyng. @ZijieZhaoMMHW & co have done a great job in making this analysis available on MATLAB, and I'm sure this will be a very useful tool in the future. But there are a few issues that I think need to be resolved before publication in JOSS. I have ticked all the review boxes that I think I can at this point, & added all the corresponding issues.
