### 2019-01-10 d3e4ac2
Update compare_M_R.md
