+++ testing/compare_M_R.md
@@ -1,78 +1,112 @@
-Comparing Outputs from **`m_mhw`** and **`RmarineHeatWaves`**
+Comparing Outputs from **`m_mhw`** and Python module.
 ==================================================================
 
-Here we provide a comparison between outputs from **`m_mhw`** and **`RmarineHeatWaves`**
+Here we provide a comparison between outputs from **`m_mhw`** and [**`marineHeatwaves.py`**](https://github.com/ecjoliver/marineHeatWaves/blob/master/marineHeatWaves.py). To simplify the testing process, we choose the `sst_full(1,2)` as the input dataset. 
 
-Let's have a look at MHW outputs in grid (1,2) during 1993 to 2016 based on climatologies and thresholds for 1982 to 2005.
+```
+sst_full=NaN(32,32,datenum(2016,12,31)-datenum(1982,1,1)+1);
+for i=1982:2016;
+    file_here=['sst_' num2str(i)];
+    load(file_here);
+    eval(['data_here=sst_' num2str(i) ';'])
+    sst_full(:,:,(datenum(i,1,1):datenum(i,12,31))-datenum(1982,1,1)+1)=data_here;
+end
+% choose data in grid (1,2)
+sst_1_2=sst_full(1,2,:);
+```
+
+We firstly write `sst_1_2` into a csv file. This is to make sure that MATLAB and python get data from the same file. We store the outputs as `MHW`,`mclim`,`m90`.
+
+```
+% write it into csv file
+csvwrite('sst_1_2.csv',[1;round(sst_1_2(:),2)]);
+% read it again to make sure MATLAB and Python get data from the same file
+sst_1_2=csvread('sst_1_2.csv');
+% detecting MHW using MATLAB
+[MHW,mclim,m90,mhw_ts]=detect(reshape(sst_1_2,1,1,length(sst_1_2)),datenum(1982,1,1):datenum(2016,12,31),datenum(1982,1,1),datenum(2005,12,31),datenum(1982,1,1),datenum(2016,12,31));
 
 ```
 
-[MHW,mclim,m90,mhw_ts]=detect(sst_full,1982,2005,1982,1993,2016);
-
-event_comp=MHW{:,:};
-event_comp=event_comp(event_comp(:,8)==1 & event_comp(:,9)==2,:);
-
+Then, we read it in python and use the marineHeatwaves.py module to analyse it. We store this result as `mhws` and `mclim`.
 
 ```
-
-There is an event during 05/10/1999 to 27/10/1999.
-
+import os
+import numpy as np
+from datetime import date
+import marineHeatWaves as mhw
+t = np.arange(date(1982,1,1).toordinal(),date(2016,12,31).toordinal()+1)
+import pandas as pd
+data = pd.read_csv('sst_1_2.csv')
+sst=[]
+for i in range(0,12784):
+    sst.append(float(np.array(data)[i]))
+sst=np.array(sst)
+mhws, clim=mhw.detect(t,sst,climatologyPeriod=[1982, 2005])
 ```
 
-event_comp(6,1:2)
+Then, we store the climatology and threshold for coming comparsion.
+
+```
+import netCDF4 as nc
+da=nc.Dataset("threshold_climatology","w",format="NETCDF4")
+da.createDimension("time",12784)
+da.createVariable("threshold","f8",("time"))
+da.createVariable("climatology","f8",("time"))
+da.variables["threshold"][:]=clim['thresh']
+da.variables['climatology'][:]=clim['seas']
+da.close()
+```
+
+We could find slight differences. Fistly, MHW (via MATLAB) contains 93 events while mhws (via python) contains 94 events. Additionally, if you carefully look at these two variables, you could find small differences in events corresponding to the same period.
+
[…]
