### 2019-01-10 b243e54
Update compare_M_R.md
