+++ testing/compare_M_R.md
@@ -97,7 +97,7 @@ Threshold are absolutely the same. So it could be caused by the fact that these 
 
 Way to detect MHWs
 -------------
-Although thresholds are different, the difference between results could also be due to different ways to detect MHW period. So we write another function called 'detect_justfortest'. This function is absolutely the same as `detect` except its threshold is one of inputs. Now, we use the `m90_p` as designed threshold, to see if MATLAB could get the same results from python. This output is stored as `MHW_p`.
+Although thresholds are different, the difference between results could also be due to different ways to detect MHW period. So we write another function called `detect_justfortest`. This function is absolutely the same as `detect` except its threshold is one of inputs. Now, we use the `m90_p` as designed threshold, to see if MATLAB could get the same results from python. This output is stored as `MHW_p`.
 
 ```
 [MHW_p,mclim,m90,mhw_ts]=detect_justfortest(reshape(sst_1_2,1,1,length(sst_1_2)),datenum(1982,1,1):datenum(2016,12,31),datenum(1982,1,1),datenum(2005,12,31),datenum(1982,1,1),datenum(2016,12,31),reshape(m90_p,1,1,366));
@@ -107,6 +107,6 @@ Now we get 94 events. Detailed comparsion has not been shown here but if you do 
 
 Different way to calculate the percentile
 -------------
-The function `prctile` in MATLAB Statistics and Machine Learning Toolbox calculates the percentile based on a particular linear regression (see [details](https://au.mathworks.com/help/stats/prctile.html?searchHighlight=prctile&s_tid=doc_srchtitle)), while Python Numpy uses another algorithm. In R's [`stats` toolbox](https://www.rdocumentation.org/packages/stats/versions/3.5.2/topics/quantile), 9 types of percentile is recorded. Numpy's is the same as Type 7 while MATLAB's is similar to Type 5. Compared to Type 7, Type 5 is more popular in the study about hydrology (Hyndman and Fan, 1996), which is just the heatwave's condition. Additionally, the original paper (Oliver et al., 2016) has not mentioned the way to calculate the percentile, and we also want to make the result consistent with other MATLAB outputs (which are mostly based on Statistics and Machine Learning Toolbox). So, after carefull discussion, we decide the keep difference and not change it.
+The function `prctile` in MATLAB Statistics and Machine Learning Toolbox calculates the percentile based on a particular linear regression (see [details](https://au.mathworks.com/help/stats/prctile.html?searchHighlight=prctile&s_tid=doc_srchtitle)), while Python Numpy uses another algorithm. In R's [`stats` toolbox](https://www.rdocumentation.org/packages/stats/versions/3.5.2/topics/quantile), 9 types of percentile is recorded. Numpy's is the same as Type 7 while MATLAB's is similar to Type 5. Compared to Type 7, Type 5 is more popular in the study about hydrology (Hyndman and Fan, 1996), which is just the heatwave's condition. Additionally, the original paper (Oliver et al., 2016) has not mentioned the way to calculate the percentile, and we also want to make the result consistent with other MATLAB outputs (which are mostly based on Statistics and Machine Learning Toolbox). So, after carefull discussion, we decide the keep difference and not change it. We also strongly suggest scientists describe their way to calculate percentile in their future especially when it is associated with methodology.
 
 
