### 2019-01-10 5bef771
Update compare_M_R.md
