### 2019-01-10 9e30a94
Update LICENCE.txt
