+++ docs/paper.bib
@@ -7,6 +7,7 @@
   pages={1090--1103},
   year={2009},
   publisher={Wiley Online Library}
+  url = {https://onlinelibrary.wiley.com/doi/abs/10.1111/j.1365-2486.2008.01823.x}
 }
 
 @article{hansen2006global,
@@ -18,6 +19,7 @@
   pages={14288--14293},
   year={2006},
   publisher={National Acad Sciences}
+  url={https://www.pnas.org/content/103/39/14288.short}
 }
 
 @article{hobday2014identification,
@@ -29,6 +31,7 @@
   pages={415--425},
   year={2014},
   publisher={Springer}
+  url={https://link.springer.com/article/10.1007/s11160-013-9326-6}
 }
 
 @article{hobday2016hierarchical,
@@ -39,6 +42,7 @@
   pages={227--238},
   year={2016},
   publisher={Elsevier}
+  url={https://www.sciencedirect.com/science/article/pii/S0079661116000057}
 }
 
 @book{pachauri2014climate,
@@ -46,6 +50,7 @@
   author={Pachauri, Rajendra K and Allen, Myles R and Barros, Vicente R and Broome, John and Cramer, Wolfgang and Christ, Renate and Church, John A and Clarke, Leon and Dahe, Qin and Dasgupta, Purnamita and others},
   year={2014},
   publisher={IPCC}
+  url={http://epic.awi.de/37530/}
 }
 
 @article{mills2013fisheries,
@@ -57,6 +62,7 @@
   pages={191--195},
   year={2013},
   publisher={JSTOR}
+  url={https://www.jstor.org/stable/24862052}
 }
 
 @misc{new2011four,
@@ -64,6 +70,7 @@
   author={New, Mark and Liverman, Diana and Schroder, Heike and Anderson, Kevin},
   year={2011},
   publisher={The Royal Society}
+  url={https://royalsocietypublishing.org/doi/full/10.1098/rsta.2010.0303}
 }
 
 @article{oliver2017unprecedented,
@@ -74,6 +81,7 @@
   pages={16101},
   year={2017},
   publisher={Nature Publishing Group}
+  url={https://www.nature.com/articles/ncomms16101}
 }
 
 @article{oliver2018longer,
@@ -85,6 +93,7 @@
   pages={1324},
   year={2018},
   publisher={Nature Publishing Group}
+  url={https://www.nature.com/articles/s41467-018-03732-9}
 }
 
 @article{oliver2018marine,
@@ -95,6 +104,7 @@
   pages={116--130},
   year={2018},
   publisher={Elsevier}
+  url={https://www.sciencedirect.com/science/article/pii/S0079661117303336}
 }
 
 @article{perkins2013measurement,
@@ -105,6 +115,7 @@
   number={13},
   pages={4500--4517},
   year={2013}
+  url={https://journals.ametsoc.org/doi/abs/10.1175/JCLI-D-12-00383.1}
 }
 
 @article{perry2005climate,
@@ -116,6 +127,7 @@
   pages={1912--1915},
   year={2005},
   publisher={American Association for the Advancement of Science}
+  url={http://science.sciencemag.org/content/308/5730/1912}
 }
 
 @article{w2018heatwaver,
@@ -125,6 +137,7 @@
   volume={3},
   pages={821},
   year={2018}
+  url={http://joss.theoj.org/papers/10.21105/joss.00821}
 }
 
 @article{schlegel2017nearshore,
@@ -135,6 +148,7 @@
   pages={189--205},
   year={2017},
   publisher={Elsevier}
+  url={https://www.sciencedirect.com/science/article/pii/S0079661116300830}
 }
 
 @article{wernberg2013extreme,
[…]
