### 2019-01-10 e9c4f6e
Update detect.m
