+++ detect.m
@@ -126,25 +126,28 @@ EventNames = {'MHW','MCS'};
 vEvent = internal.stats.getParamVal(vEvent,EventNames,...
     '''Event''');
 
+%%  "What if cli_start-window or cli_end+window exceeds the time range of data"
+
 ahead_date=time(1)-(cli_start-vWindowHalfWidth);
 after_date=cli_end+vWindowHalfWidth-time(end);
 temp_clim=ClimTemp(:,:,ClimTime>=cli_start-vWindowHalfWidth & ClimTime<=cli_end+vWindowHalfWidth);
 
-if ahead_date>0 && after_date>0;
+if ahead_date>0 && after_date>0
     temp_clim=cat(3,NaN(size(temp_clim,1),size(temp_clim,2),ahead_date), ...
     temp_clim,NaN(size(temp_clim,1),size(temp_clim,2),after_date));
-elseif ahead_date>0 && after_date<=0;
+elseif ahead_date>0 && after_date<=0
     temp_clim=cat(3,NaN(size(temp_clim,1),size(temp_clim,2),ahead_date), ...
     temp_clim);
-elseif ahead_date<=0 && after_date>0;
+elseif ahead_date<=0 && after_date>0
         temp_clim=cat(3, ...
             temp_clim,NaN(size(temp_clim,1),size(temp_clim,2),after_date));
 else
-    temp_clim=temp_clim;
+    
 end
 
 temp_mhw=temp(:,:,time>=mhw_start & time<=mhw_end);
 
+%% Calculating climatology and thresholds
 
 date_true=datevec(cli_start-vWindowHalfWidth:cli_end+vWindowHalfWidth);
 date_true=date_true(:,1:3);
@@ -160,7 +163,7 @@ m90=NaN(size(temp,1),size(temp,2),366);
 
 for i=1:366
     if i == 60
-        ;
+        
     else
         ind_fake=ind;
         ind_fake(fake_doy==i & ~ismember(datenum(date_true),cli_start:cli_end))=nan;
@@ -168,22 +171,18 @@ for i=1:366
     m90(:,:,i) = quantile(temp_clim(:,:,any(ind_fake'>=(ind_fake(fake_doy == i)-vWindowHalfWidth) & ind_fake' <= (ind_fake(fake_doy ==i)+vWindowHalfWidth),2)),vThreshold,3);
     mclim(:,:,i) = mean(temp_clim(:,:,any(ind_fake'>=(ind_fake(fake_doy == i)-vWindowHalfWidth) & ind_fake' <= (ind_fake(fake_doy ==i)+vWindowHalfWidth),2)),3,'omitnan');
     end
-end   
+end
+% Dealing with Feb29
 m90(:,:,60) = mean(m90(:,:,[59 61]),3,'omitnan');
 mclim(:,:,60) = mean(mclim(:,:,[59 61]),3,'omitnan');
 
-
-true_time = datevec(time);
-true_time = true_time(:,1:3);
-true_time(:,1) = 2012;
-time_doy = day(datetime(true_time),'dayofyear');
-
+% Does running averages of threshold and clim..
 
 m90long=smoothdata(cat(3,m90,m90,m90),3,'movmean',vsmoothPercentileWidth);
 m90=m90long(:,:,367:367+365);
 mclimlong=smoothdata(cat(3,mclim,mclim,mclim),3,'movmean',vsmoothPercentileWidth);
 mclim=mclimlong(:,:,367:367+365);
-% does running averages of threshold and clim..
+
 
[…]
