### 2019-01-10 72b8a82
Update compare_M_R.md
