+++ docs/paper.bib
@@ -7,7 +7,7 @@
   pages={1090--1103},
   year={2009},
   publisher={Wiley Online Library}
-  url = {https://onlinelibrary.wiley.com/doi/abs/10.1111/j.1365-2486.2008.01823.x}
+  url={https://onlinelibrary.wiley.com/doi/abs/10.1111/j.1365-2486.2008.01823.x}
 }
 
 @article{hansen2006global,
