+++ docs/paper.bib
@@ -1,12 +1,13 @@
 @article{garrabou2009mass,
   title={Mass mortality in Northwestern Mediterranean rocky benthic communities: effects of the 2003 heat wave},
   author={Garrabou, Joaquim and Coma, Rafael and Bensoussan, Nathaniel and Bally, Marc and Chevaldonn{\'e}, Pierre and Cigliano, M and D{\'\i}az, David and Harmelin, Jean-Georges and Gambi, Mar{\'\i}a Cristina and Kersting, DK and others},
+  doi={ 10.1111/j.1365-2486.2008.01823.x},
   journal={Global change biology},
   volume={15},
   number={5},
   pages={1090--1103},
   year={2009},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
   url={https://onlinelibrary.wiley.com/doi/abs/10.1111/j.1365-2486.2008.01823.x}
 }
 
@@ -18,7 +19,7 @@
   number={39},
   pages={14288--14293},
   year={2006},
-  publisher={National Acad Sciences}
+  publisher={National Acad Sciences},
   url={https://www.pnas.org/content/103/39/14288.short}
 }
 
@@ -30,7 +31,7 @@
   number={2},
   pages={415--425},
   year={2014},
-  publisher={Springer}
+  publisher={Springer},
   url={https://link.springer.com/article/10.1007/s11160-013-9326-6}
 }
 
@@ -41,7 +42,7 @@
   volume={141},
   pages={227--238},
   year={2016},
-  publisher={Elsevier}
+  publisher={Elsevier},
   url={https://www.sciencedirect.com/science/article/pii/S0079661116000057}
 }
 
@@ -49,7 +50,7 @@
   title={Climate change 2014: synthesis report. Contribution of Working Groups I, II and III to the fifth assessment report of the Intergovernmental Panel on Climate Change},
   author={Pachauri, Rajendra K and Allen, Myles R and Barros, Vicente R and Broome, John and Cramer, Wolfgang and Christ, Renate and Church, John A and Clarke, Leon and Dahe, Qin and Dasgupta, Purnamita and others},
   year={2014},
-  publisher={IPCC}
+  publisher={IPCC},
   url={http://epic.awi.de/37530/}
 }
 
@@ -61,7 +62,7 @@
   number={2},
   pages={191--195},
   year={2013},
-  publisher={JSTOR}
+  publisher={JSTOR},
   url={https://www.jstor.org/stable/24862052}
 }
 
@@ -69,7 +70,7 @@
   title={Four degrees and beyond: the potential for a global temperature increase of four degrees and its implications},
   author={New, Mark and Liverman, Diana and Schroder, Heike and Anderson, Kevin},
   year={2011},
-  publisher={The Royal Society}
+  publisher={The Royal Society},
   url={https://royalsocietypublishing.org/doi/full/10.1098/rsta.2010.0303}
 }
 
@@ -80,7 +81,7 @@
   volume={8},
   pages={16101},
   year={2017},
-  publisher={Nature Publishing Group}
+  publisher={Nature Publishing Group},
   url={https://www.nature.com/articles/ncomms16101}
 }
 
@@ -92,7 +93,7 @@
   number={1},
   pages={1324},
   year={2018},
-  publisher={Nature Publishing Group}
+  publisher={Nature Publishing Group},
   url={https://www.nature.com/articles/s41467-018-03732-9}
 }
 
@@ -103,7 +104,7 @@
   volume={161},
   pages={116--130},
   year={2018},
-  publisher={Elsevier}
+  publisher={Elsevier},
   url={https://www.sciencedirect.com/science/article/pii/S0079661117303336}
 }
 
[…]
