### 2019-01-10 bd5188b
Update paper.bib
