+++ docs/paper.bib
@@ -1,7 +1,7 @@
 @article{garrabou2009mass,
   title={Mass mortality in Northwestern Mediterranean rocky benthic communities: effects of the 2003 heat wave},
   author={Garrabou, Joaquim and Coma, Rafael and Bensoussan, Nathaniel and Bally, Marc and Chevaldonn{\'e}, Pierre and Cigliano, M and D{\'\i}az, David and Harmelin, Jean-Georges and Gambi, Mar{\'\i}a Cristina and Kersting, DK and others},
-  doi={ 10.1111/j.1365-2486.2008.01823.x},
+  doi={10.1111/j.1365-2486.2008.01823.x},
   journal={Global change biology},
   volume={15},
   number={5},
@@ -14,6 +14,7 @@
 @article{hansen2006global,
   title={Global temperature change},
   author={Hansen, James and Sato, Makiko and Ruedy, Reto and Lo, Ken and Lea, David W and Medina-Elizade, Martin},
+  doi={10.1073/pnas.0606291103},
   journal={Proceedings of the National Academy of Sciences},
   volume={103},
   number={39},
@@ -26,6 +27,7 @@
 @article{hobday2014identification,
   title={Identification of global marine hotspots: sentinels for change and vanguards for adaptation action},
   author={Hobday, Alistair J and Pecl, Gretta T},
+  doi={10.1007/s11160-013-9326-6},
   journal={Reviews in Fish Biology and Fisheries},
   volume={24},
   number={2},
@@ -38,6 +40,7 @@
 @article{hobday2016hierarchical,
   title={A hierarchical approach to defining marine heatwaves},
   author={Hobday, Alistair J and Alexander, Lisa V and Perkins, Sarah E and Smale, Dan A and Straub, Sandra C and Oliver, Eric CJ and Benthuysen, Jessica A and Burrows, Michael T and Donat, Markus G and Feng, Ming and others},
+  doi={10.1016/j.pocean.2015.12.014},
   journal={Progress in Oceanography},
   volume={141},
   pages={227--238},
@@ -57,6 +60,7 @@
 @article{mills2013fisheries,
   title={Fisheries management in a changing climate: lessons from the 2012 ocean heat wave in the Northwest Atlantic},
   author={Mills, Katherine E and Pershing, Andrew J and Brown, Curtis J and Chen, Yong and Chiang, Fu-Sung and Holland, Daniel S and Lehuta, Sigrid and Nye, Janet A and Sun, Jenny C and Thomas, Andrew C and others},
+  doi={10.5670/oceanog.2013.27},
   journal={Oceanography},
   volume={26},
   number={2},
@@ -69,6 +73,7 @@
 @misc{new2011four,
   title={Four degrees and beyond: the potential for a global temperature increase of four degrees and its implications},
   author={New, Mark and Liverman, Diana and Schroder, Heike and Anderson, Kevin},
+  doi={10.1098/rsta.2010.0303},
   year={2011},
   publisher={The Royal Society},
   url={https://royalsocietypublishing.org/doi/full/10.1098/rsta.2010.0303}
@@ -77,6 +82,7 @@
 @article{oliver2017unprecedented,
   title={The unprecedented 2015/16 Tasman Sea marine heatwave},
   author={Oliver, Eric CJ and Benthuysen, Jessica A and Bindoff, Nathaniel L and Hobday, Alistair J and Holbrook, Neil J and Mundy, Craig N and Perkins-Kirkpatrick, Sarah E},
+  doi={10.1038/ncomms16101},
   journal={Nature communications},
   volume={8},
   pages={16101},
@@ -88,6 +94,7 @@
 @article{oliver2018longer,
[…]
