### 2019-01-10 ZijieZhaoMMHW/m_mhw1.0#6 @ZijieZhaoMMHW
Please see https://github.com/ZijieZhaoMMHW/m_mhw1.0/blob/master/testing/compare_M_R.md
