### 2019-01-10 #1124 @ZijieZhaoMMHW
Hi @mvdh7  I have responded all issues. Please see if you are ok with it.
