### 2019-01-10 691d0e7
Update compare_M_R.md
