### 2019-01-14 ZijieZhaoMMHW/m_mhw1.0#1 @mvdh7
It appears that you have either decided to credit me as the author of your software, or you have copied and pasted a license from one of my repositories. This is an odd decision as none of my code is published with JOSS. I suggest you update the license so that it applies to your own software and check it matches the OSI approved list as required by JOSS.
