### 2019-01-14 24a16b0
Update LICENCE.txt
