### 2019-01-14 ZijieZhaoMMHW/m_mhw1.0#1 @ZijieZhaoMMHW
Please see https://github.com/ZijieZhaoMMHW/m_mhw1.0/blob/master/docs/LICENCE.txt.

Now I use the GNU GENERAL PUBLIC LICENSE Version 3.
