### 2019-01-14 ZijieZhaoMMHW/m_mhw1.0#5 @mvdh7
My understanding of the JOSS review process is that it's supposed to help improve code not just check that it works. The comments below are given in that spirit, but I am quite happy for them to be ignored if I have misunderstood, and if if the Editor feels these are not appropriate or necessary for publication.

Eliminating the warning flags would be a useful exercise especially if combined with improvements to readability and style throughout. The code is not easy to follow for several reasons: (1) there is a lack of comments within the functions describing what is going on; (2) some basic good practices are ignored, such as spaces around `=` characters; (3) there are some very long lines that should be broken up with `...`.

1. The majority of your warning flags are due to unnecessary semi-colons, usually at the end of a for/if/while statement. While these don't cause any problems, they do make it much more difficult to spot other, more useful warnings. They require very little effort to remove.

2. You then have some unnecessary `find` statements, of the format:

`vector_variable(find(logical_expression))`

These work equally well - and marginally faster - if you simply put:

`vector_variable(logical_expression)`

3. You have some variables that are assigned within loops that are never used, which can simply be eliminated to avoid wasting memory.

4. Based on your reply above, I think you may be mixing up "parallel" with "preallocation". MATLAB does not suggest parallel computation in your code, it suggests that you preallocate some variables in your loops. This means that if you are assigning values into a vector one by one in a loop, you should ideally create an empty vector before the loop begins, of the correct size, into which the results are added.

For example, part of your code in `detect()` looks something like this:

```matlab
mhwint_max = [];
for le = 1:size(potential_event,1)
    maxanom = <some code>;
    mhwint_max = [mhwint_max; maxanom];
end
```

To preallocate you would change this to:

```matlab
mhwint_max = NaN(size(potential_event,1),1);
for le = 1:size(potential_event,1)
    maxanom = <some code>;
    mhwint_max(le) = maxanom;
end
```

This allows your computer to use memory more efficiently and your code to run faster.

5. One warning flag simply tells you it is faster to use `str2double` instead of `str2num`. This is an extremely straightforward switch. Why resist?
