### 2019-01-14 ZijieZhaoMMHW/m_mhw1.0#1 @mvdh7
There are still lines in the license file that need to be updated with your details
e.g. line 656 `<program>  Copyright (C) <year>  <name of author>`
And others! :)
