### 2019-01-14 ZijieZhaoMMHW/m_mhw1.0#3 @mvdh7
Thank you.

I should have noticed this previously but clearly did not spot it. Is there also documentation for the `event_line()` and `mean_and_trend()` function input/outputs too? Are these intended for use by the end user?
