### 2019-01-14 578367a
Update detect.m
