+++ detect.m
@@ -192,18 +192,12 @@ date_mhw=datevec(mhw_start:mhw_end);
 date_mhw(:,1)=2000;
 indextocal = day(datetime(date_mhw),'dayofyear');
 
-ts=str2num(datestr(mhw_start:mhw_end,'YYYYmmdd'));
+ts=str2double(string(datestr(mhw_start:mhw_end,'YYYYmmdd')));
 
 mhw_ts=zeros(x_size,y_size,length(ts));
 
 
-mhwstart=[];
-mhwend=[];
-mduration=[];
-mhwint_max=[];
-mhwint_mean=[];
-mhwint_var=[];
-mhwint_cum=[];
+
 
 MHW=[];
 
@@ -241,27 +235,35 @@ switch vEvent
                         end
                     end
                     
-                    if ~isempty(potential_event);
+                    if ~isempty(potential_event)
                         
                         potential_event=potential_event((potential_event(:,2)-potential_event(:,1)+1)>=vminDuration,:);
                         
-                        if ~isempty(potential_event);
+                        if ~isempty(potential_event)
                             
                             gaps=(potential_event(2:end,1) - potential_event(1:(end-1),2) - 1);
                             
-                            while min(gaps)<=vmaxGap;
-%                                  potential_event(find(gaps<=vmaxGap),2)=potential_event(find(gaps<=vmaxGap)+1,2);
-%                                  potential_event(find(gaps<=vmaxGap)+1,:)=[];
-%                                  gaps=(potential_event(2:end,1) - potential_event(1:(end-1),2) - 1);
-                                 
-                                 potential_event(find(gaps<=vmaxGap),2)=potential_event(find(gaps<=vmaxGap)+1,2);
-                                 loc_should_del=(find(gaps<=vmaxGap)+1);
-                                 loc_should_del=loc_should_del(~ismember(loc_should_del,find(gaps<=vmaxGap)));
-                                 potential_event(loc_should_del,:)=[];
-                                 gaps=(potential_event(2:end,1) - potential_event(1:(end-1),2) - 1);
+                            while min(gaps)<=vmaxGap
+                                %                                  potential_event(find(gaps<=vmaxGap),2)=potential_event(find(gaps<=vmaxGap)+1,2);
+                                %                                  potential_event(find(gaps<=vmaxGap)+1,:)=[];
+                                %                                  gaps=(potential_event(2:end,1) - potential_event(1:(end-1),2) - 1);
+                                
+                                potential_event(find(gaps<=vmaxGap),2)=potential_event(find(gaps<=vmaxGap)+1,2);
+                                loc_should_del=(find(gaps<=vmaxGap)+1);
+                                loc_should_del=loc_should_del(~ismember(loc_should_del,find(gaps<=vmaxGap)));
+                                potential_event(loc_should_del,:)=[];
+                                gaps=(potential_event(2:end,1) - potential_event(1:(end-1),2) - 1);
                             end
                             
-                            for le=1:size(potential_event,1);
+                            mhwstart=NaN(size(potential_event,1),1);
+                            mhwend=NaN(size(potential_event,1),1);
+                            mduration=NaN(size(potential_event,1),1);
+                            mhwint_max=NaN(size(potential_event,1),1);
+                            mhwint_mean=NaN(size(potential_event,1),1);
+                            mhwint_var=NaN(size(potential_event,1),1);
+                            mhwint_cum=NaN(size(potential_event,1),1);
+                            
+                            for le=1:size(potential_event,1)
                                 event_here=potential_event(le,:);
                                 endtime=ts(event_here(2));
                                 starttime=ts(event_here(1));
@@ -270,33 +272,28 @@ switch vEvent
                                 manom=mrow-mcl;
[…]
