### 2019-01-14 ZijieZhaoMMHW/m_mhw1.0#5 @ZijieZhaoMMHW
Thanks a lot for careful and thoughtful review. I have edited the script for your suggestion. Thanks a lot. Now most warning flag has been removed.

I need to mention that `find` still needs to be used in `potential_event(find(gaps<=vmaxGap),2)=potential_event(find(gaps<=vmaxGap)+1,2);` due to the fact that `gaps` has different size from `potential_event`, here I just use `find` to determine the location that I want to fill in so I have to use `find`, but MATLAB initially think `gaps` has the same size with `potential_event` so it suggests no `find`.
Please see https://github.com/ZijieZhaoMMHW/m_mhw1.0/blob/master/detect.m.
