### 2019-01-14 3c3d4e9
Update LICENCE.txt
