### 2019-01-14 672fc4c
Update LICENCE.txt
