+++ docs/LICENCE.txt
@@ -633,7 +633,7 @@ state the exclusion of warranty; and each file should have at least
 the "copyright" line and a pointer to where the full notice is found.
 
     <A MATLAB toolbox to detect and analyze marine heatwaves>
-    Copyright (C) <2019r>  <Zijie Zhao>
+    Copyright (C) <2019>  <Zijie Zhao>
 
     This program is free software: you can redistribute it and/or modify
     it under the terms of the GNU General Public License as published by
