### 2019-01-14 b122b40
Update compare_M_R.md
