+++ testing/compare_M_R.md
@@ -109,6 +109,6 @@ Now we get 94 events. Detailed comparsion has not been shown here but if you do 
 
 Different way to calculate the percentile
 -------------
-The function `prctile` in MATLAB Statistics and Machine Learning Toolbox calculates the percentile based on a particular linear regression (see [details](https://au.mathworks.com/help/stats/prctile.html?searchHighlight=prctile&s_tid=doc_srchtitle)), while Python Numpy uses another algorithm. In R's [`stats` toolbox](https://www.rdocumentation.org/packages/stats/versions/3.5.2/topics/quantile), 9 types of percentile is recorded. Numpy's is the same as Type 7 while MATLAB's is similar to Type 5. Compared to Type 7, Type 5 is more popular in the study about hydrology (Hyndman and Fan, 1996), which is just the heatwave's condition. Additionally, the original paper (Oliver et al., 2016) has not mentioned the way to calculate the percentile, and we also want to make the result consistent with other MATLAB outputs (which are mostly based on Statistics and Machine Learning Toolbox). So, after carefull discussion, we decide the keep difference and not change it. We also strongly suggest scientists to describe their way to calculate percentile in their future research especially when it is associated with methodology.
+The function `prctile` in MATLAB Statistics and Machine Learning Toolbox calculates the percentile based on a particular linear regression (see [details](https://au.mathworks.com/help/stats/prctile.html?searchHighlight=prctile&s_tid=doc_srchtitle)), while Python Numpy uses another algorithm. In R's [`stats` toolbox](https://www.rdocumentation.org/packages/stats/versions/3.5.2/topics/quantile), 9 types of percentile is recorded. Numpy's is the same as Type 7 while MATLAB's is similar to Type 5. Compared to Type 7, Type 5 is more popular in the study about hydrology (Hyndman and Fan, 1996), which is just the heatwave's condition. Additionally, the original paper (Oliver et al., 2016) has not mentioned the way to calculate the percentile, and we also want to make the result consistent with other MATLAB outputs (which are mostly based on Statistics and Machine Learning Toolbox). So, after carefull discussion, we decide to keep the difference and not change it. We also strongly suggest scientists to describe their way to calculate percentile in their future research especially when it is associated with methodology.
 
 
