### 2019-01-14 ZijieZhaoMMHW/m_mhw1.0#1 @ZijieZhaoMMHW
Hi. I have modified the license.

see Line 656 `<m_mhw>  Copyright (C) <2019>  <Zijie Zhao>`
see Line 635 `<A MATLAB toolbox to detect and analyze marine heatwaves>
    Copyright (C) <2019>  <Zijie Zhao>`
