### 2019-01-14 ZijieZhaoMMHW/m_mhw1.0#3 @ZijieZhaoMMHW
Sincere thanks for suggestion. But considering adding description of them into readme.md would make it too long and they are well documented in MATLAB `help` text, so I just mention this point in readme.md. See https://github.com/ZijieZhaoMMHW/m_mhw1.0/blob/master/README.md
