### 2019-01-21 34f2747
Delete event_line.m
