### 2019-01-21 6af2359
Delete mean_and_trend.m
