+++ mean_and_trend.m
@@ -1,374 +1,0 @@
-function [mean_metric,annual_metric,trend_metric,p_metric]=mean_and_trend(MHW,mhw_ts,data_start,varargin);
-%mean_and_trend - calculating mean states and trends of event metrics
-%  Syntax
-%  [mean_metric,annual_metric,trend_metric,p_metric]=mean_and_trend(MHW,mhw_ts,data_start);
-%  [mean_metric,annual_metric,trend_metric,p_metric]=mean_and_trend(MHW,mhw_ts,data_start,'Metric','Duration');
-%  
-%  Description
-%  [mean_metric,annual_metric,trend_metric,p_metric]=mean_and_trend(MHW,mhw_ts,data_start)
-%  returns the mean states and annual trends of MHW/MCS frequency (number
-%  of events per year) based on event table MHW and MHW/MCS time series
-%  MHW_TS. The start year (DATA_START) is the first year of MHW_TS.
-%
-%  [mean_metric,annual_metric,trend_metric,p_metric]=mean_and_trend(MHW,mhw_ts,data_start,'Metric','Duration')
-%  returns the mean states and annual trends of MHW/MCS duration. 
-%
-%  Input Arguments
-%
-%   MHW, mhw_ts - Outputs from function detect
-%
-%   data_start - A numeric value indicating the start year of MHW_TS.
-%
-%   'Metric' - Default is 'Frequency'. The metric for which mean states and annual trends are
-%   calculated.
-%            'Frequency' - The annual number of events.
-%            'Duration' - The duration of events.
-%            'MaxInt' - The maximum intensity of events.
-%            'MeanInt' - The mean intensity of events.
-%            'CumInt' - The cumulative intensity of events.
-%            'Days' - The annual total MHW/MCS days.
-%
-%  Output Arguments
-%   
-%   mean_metric - A 2D numeric matrix (m-by-n) containing the mean states
-%   of MHW/MCS metrics.
-%   
-%   annual_metric - A 3D numeric matrix (m-by-n-by-Y) containing annual
-%   mean MHW/MCS metrics, whre Y indicates the number of year based on the
-%   start year DATA_START and the size of MHW_TS.
-%
-%   trend_metric - A 2D numeric matrix (m-by-n) containing linear trend
-%   calculated from ANNUAL_METRIC in unit of 'unit of metric/year'.
-%
-%   p_metric - A 2D numeric matrix (m-by-n) containing p value of
-%   TREND_METRIC.
-
-
-paramNames = {'Metric'};
-defaults   = {'Frequency'};
-
-[vMetric]...
-    = internal.stats.parseArgs(paramNames, defaults, varargin{:});
-
-MetricNames = {'Frequency','Duration','MaxInt','MeanInt','CumInt','Days'};
-vMetric = internal.stats.getParamVal(vMetric,MetricNames,...
-    '''Metric''');
-
-[x,y,t]=size(mhw_ts);
-
-switch vMetric
-    case 'Duration'
-        
-        MHW=MHW{:,:};
-        full_mhw_start=datevec(num2str(MHW(:,1)),'yyyymmdd');
-        full_mhw_end=datevec(num2str(MHW(:,2)),'yyyymmdd');
-        metric=MHW(:,3);
-        
-        period_used=datenum(data_start,1,1):(datenum(data_start,1,1)+size(mhw_ts,3)-1);
-        period_used=datevec(period_used);
-        
-        years_mhw=unique(period_used(:,1),'rows');
-        
-        mean_metric=NaN(x,y);
-        trend_metric=NaN(x,y);
-        p_metric=NaN(x,y);
-        annual_metric=NaN(x,y,length(years_mhw));
-        
[…]
