### 2019-01-21 daaa4ed
Delete detect_justfortest.m
