+++ testing/detect_justfortest.m
@@ -1,292 +1,0 @@
-function [MHW,mclim,m90,mhw_ts]=detect_justfortest(temp,time,cli_start,cli_end,mhw_start,mhw_end,m90,varargin)
-% This function is designed for test.
-
-paramNames = {'Event','Threshold','WindowHalfWidth','smoothPercentileWidth','minDuration',...
-    'maxGap','ClimTemp','ClimTime'};
-defaults   = {'MHW',0.9,5,31,5,2,temp,time};
-% 
-% varargin = reshape(varargin,2,length(varargin)/2);
-% 
-% for i = 1:length(defaults)
-%     if any(ismember(varargin(1,:),paramNames{i}))
-%        feval(@()assignin('caller',paramNames{i},varargin{2,ismember(varargin(1,:),paramNames{i})}))
-%     end      
-% end
-
-
-[vEvent, vThreshold,vWindowHalfWidth,vsmoothPercentileWidth,vminDuration,vmaxGap,ClimTemp,ClimTime]...
-    = internal.stats.parseArgs(paramNames, defaults, varargin{:});
-
-EventNames = {'MHW','MCS'};
-vEvent = internal.stats.getParamVal(vEvent,EventNames,...
-    '''Event''');
-
-ahead_date=time(1)-(cli_start-vWindowHalfWidth);
-after_date=cli_end+vWindowHalfWidth-time(end);
-temp_clim=ClimTemp(:,:,ClimTime>=cli_start-vWindowHalfWidth & ClimTime<=cli_end+vWindowHalfWidth);
-
-if ahead_date>0 && after_date>0;
-    temp_clim=cat(3,NaN(size(temp_clim,1),size(temp_clim,2),ahead_date), ...
-    temp_clim,NaN(size(temp_clim,1),size(temp_clim,2),after_date));
-elseif ahead_date>0 && after_date<=0;
-    temp_clim=cat(3,NaN(size(temp_clim,1),size(temp_clim,2),ahead_date), ...
-    temp_clim);
-elseif ahead_date<=0 && after_date>0;
-        temp_clim=cat(3, ...
-            temp_clim,NaN(size(temp_clim,1),size(temp_clim,2),after_date));
-else
-    temp_clim=temp_clim;
-end
-
-temp_mhw=temp(:,:,time>=mhw_start & time<=mhw_end);
-
-
-date_true=datevec(cli_start-vWindowHalfWidth:cli_end+vWindowHalfWidth);
-date_true=date_true(:,1:3);
-
-date_false = date_true;
-date_false(:,1) = 2012;
-
-fake_doy = day(datetime(date_false),'dayofyear');
-ind = 1:length(date_false);
-
-mclim=NaN(size(temp,1),size(temp,2),366);
-% m90=NaN(size(temp,1),size(temp,2),366);
-
-for i=1:366
-    if i == 60
-        ;
-    else
-        ind_fake=ind;
-        ind_fake(fake_doy==i & ~ismember(datenum(date_true),cli_start:cli_end))=nan;
-        
-%     m90(:,:,i) = prctile_python_R(temp_clim(:,:,any(ind_fake'>=(ind_fake(fake_doy == i)-vWindowHalfWidth) & ind_fake' <= (ind_fake(fake_doy ==i)+vWindowHalfWidth),2)),vThreshold);
-    mclim(:,:,i) = mean(temp_clim(:,:,any(ind_fake'>=(ind_fake(fake_doy == i)-vWindowHalfWidth) & ind_fake' <= (ind_fake(fake_doy ==i)+vWindowHalfWidth),2)),3,'omitnan');
-    end
[…]
