### 2019-01-21 e8f6c8e
Add Codacy badge
