### 2019-01-21 ZijieZhaoMMHW/m_mhw1.0#5 @ZijieZhaoMMHW
Hi @mvdh7. for your consideration, I send the toolbox to @codacy-badger and get the code quality A. Please see https://github.com/ZijieZhaoMMHW/m_mhw1.0
