### 2019-01-21 ZijieZhaoMMHW/m_mhw1.0#5 @mvdh7
Nice work. I hope this has helped improve the code.

Thank you for pointing out the code quality tester - I hadn't come across that before. However, I don't think it supports MATLAB, so the badge it has given may not be very meaningful.
