### 2019-01-21 #1124 @arfon
@ZijieZhaoMMHW - could you please rename your `LICENCE.txt` to `LICENSE.txt` and move it to the root (top level) of your repository please?
